+++ packages/discovery-react-components/src/utils/document/__tests__/documentUtils.spec.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { render } from '@testing-library/react';
-import { uniqRects, findOffsetInDOM, getTextNodeAndOffset } from '../documentUtils';
+import { uniqRects, findOffsetInDOM, getTextNodeAndOffset, spansIntersect } from '../documentUtils';
 import { data as textNodesData, watsonIndex } from '../__fixtures__/textNodeData';
 
 describe('uniqRects', () => {
@@ -153,7 +153,9 @@ describe('findOffsetInDOM', () => {
     expect(beginOffset).toEqual(734);
     expect(endOffset).toEqual(107);
   });
+});
 
+describe('getTextNodeAndOffset', () => {
   it('finds offset when text spans multiple TextNodes', () => {
     const MAX_TEXT_NODE_LEN = 65536;
 
@@ -175,5 +177,97 @@ describe('findOffsetInDOM', () => {
     // expect that offset will be within 2nd text node, therefore subtract 1st text node length
     const firstNode = node.childNodes[0] as Node;
     expect(textOffset).toEqual(watsonIndex - firstNode.textContent!.length);
+  });
+
+  it('calculates proper offset with encoded text', () => {
+    const { container } = render(
+      <div
+        data-child-begin={123}
+        data-child-end={196}
+        data-orig-text="Maecenas convallis neque id elit laoreet &amp;amp; quis aliquam velit aliquam."
+      >
+        Maecenas convallis neque id elit laoreet &amp; quis aliquam velit aliquam.
+      </div>
+    );
+
+    const node = container.querySelector('[data-child-begin]');
+    const { textNode, textOffset } = getTextNodeAndOffset(node as HTMLElement, 183);
+
+    // text content in DOM should be the "decoded" text
+    expect(textNode.textContent).toEqual(
+      'Maecenas convallis neque id elit laoreet & quis aliquam velit aliquam.'
+    );
+    // while the offset passed in to func is +60, once we account for decoded text in DOM, the offset should be +56
+    expect(textOffset).toEqual(56);
+  });
+});
+
+describe('spansIntersect', () => {
+  it('checks that two cells intersect', () => {
+    const cell1 = {
+      start_offset: 138812,
+      end_offset: 139245,
+      field: 'text'
+    };
+
+    const cell2 = {
+      start_offset: 139244,
+      end_offset: 139300,
+      field: 'text'
+    };
+
+    const cell3 = {
+      start_offset: 139248,
+      end_offset: 139300,
+      field: 'text'
+    };
+
+    let result = spansIntersect(
+      { begin: cell1.start_offset, end: cell1.end_offset },
+      { begin: cell2.start_offset, end: cell2.end_offset }
+    );
+
+    expect(result).toBeTruthy();
+
+    result = spansIntersect(
+      { begin: cell1.start_offset, end: cell1.end_offset },
+      { begin: cell3.start_offset, end: cell3.end_offset }
+    );
+
+    expect(result).toBeFalsy();
+  });
+
+  it('checks that the end intersections are exclusive', () => {
+    const cell1 = {
+      start_offset: 138812,
+      end_offset: 139245,
+      field: 'text'
+    };
+
+    const cell2 = {
+      start_offset: 139245,
+      end_offset: 139300,
+      field: 'text'
+    };
+
+    const cell3 = {
+      start_offset: 139244,
+      end_offset: 139300,
+      field: 'text'
+    };
+
+    let result = spansIntersect(
+      { begin: cell1.start_offset, end: cell1.end_offset },
+      { begin: cell2.start_offset, end: cell2.end_offset }
+    );
+
+    expect(result).toBeFalsy();
+
+    result = spansIntersect(
+      { begin: cell1.start_offset, end: cell1.end_offset },
+      { begin: cell3.start_offset, end: cell3.end_offset }
+    );
+
+    expect(result).toBeTruthy();
   });
 });